requirements.txt
asgiref==3.8.1
DateTime==5.5
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
setuptools==80.8.0
sqlparse==0.5.3
uritemplate==4.1.1
zope.interface==7.2
